Bouncy Ball Bonanza
Added a Bouncy Ball to the Starting Room (turn to your right & it's in the back corner).
It also has color variants (Currently Baja, Black, Blue, Brown, and Red).
My plan over the winter break is to add new objects to the game, place them in other rooms, and add new rooms. I also want to add new color variations to Zombert as well, but I may also add color variants to the objects to so that they can spice up the levels.
Everything but the Kitchen Sink usually means a large amount of stuff, so it's time I start adding more stuff and adding more color variants of said stuff.